+++ projects/sip-alain/src/sip/services/sip-rest.service.ts
@@ -194,9 +194,6 @@ export class SipRestService {
                 }));
             }
 
-            if ('$showPreLoad' in owner)
-                http = owner.$showPreLoad(http);
-
             return http;
         } else
             return http;