+++ src/app/routes/ui-demo/ui-demo-shared/components/form/form.component.ts
@@ -12,7 +12,6 @@ export class FormComponent extends SipComponent {
 
     constructor(vcf: ViewContainerRef) {
         super(vcf);
-        console.log(this)
     }
 
     @SipFormGroup((target: FormComponent) => {