+++ projects/sip-alain/CHANGELOG.md
@@ -6,6 +6,7 @@
 * OPT: ISipFormGroupParams<T>
 * OPT: rTab.closable
 * Add: SipFormGroup 验证不通过可以定义提示信息
+* 去掉自动loading, SipOnShow
 
 ## 0.5.0